cracion de vista y controlador de transactions
